ci(deploy-to-kind-cluster): added default env vars (#5473)
* ci(deploy-to-kind-cluster): added default env vars

\# Description
* classify between `darwin` vs. `linux` systems when running `ci/deploy-to-kind-cluster.sh`
* default `KUBE2E_TESTS` to `eds`, rather than nothing

This commit can be used to run `ci/deploy-to-kind-cluster.sh` on a mac off of a fresh clone with no configuration

\# Context
One of the more common use-cases for `deploy-to-kind-cluster` is in local development.  _Currently_, when running the script, some tweaks are needed to have it _succeed_ on a Mac.  This is a bit painful, and is time that not every person on a Mac should need to figure out in order to test locally.