feat(apiserver): allow ClusterRole --> Role (#3861)
There has been a customer ask to [Make it optional for Gloo Edge (Fed) to scan for Deployment and K8s core CRD](https://github.com/solo-io/gloo/issues/6293).  Roughly, this translates to a tightening of security restrictions.

This PR addresses this via the introduction of the `gloo-fed.glooFedApiserver.namespaceRestrictedMode` helm parameter.  When set, it modifies the objects present in `glooo-ee-rbac.yaml` to operate based on single-namespace permissions.  This breaks (some) functionality of `gloo-fed-ui`, but :shrug:, as it _was_ a direct customer ask.  It's opt-in, so no harm, no foul.