6983 http protocol options (#7403) (#7412)
* 6983 http protocol options (#7403)

* Fix HTTP protocol options issue and update unit tests

The core issue here is that we are guarding against users providing http1 options and simultaneously configuring the upstream to use http2. The original code had a simple guard `in.GetConnectionConfig() != nil` to protect this. However, this was too broad as there were other options in the `ConnectionConfig` object that should have been allowed to be specified (and weren't). Instead, it is only the `ConnectionConfig.Http1ProtocolOptions` object that has these protocol-specific optionst that we want to prevent users from enabling when http2 is selected. This fix updates the statement to use this more specific condition.

I also found that the unit tests were a bit messy here. The `createTestUpstreamWithProtocolOptions` function was updated to be a bit more direct and less abstract in the object it creates. We drop the boolean `includeProtocolSelection` boolean - it defaults to 0 anyways (`USE_CONFIGURED_PROTOCOL`), so why not just use the existing `protocolSelection` parameter to that function instead? Additionally, create parameters to set `useHttp2` and `Http1ProtocolOptions`. The tests were updated according to the new signature, and this signature makes the deny condition much easier to test for.

* Change createTestUpstreamWithProtocolOptions to use a Http1ProtocolOptions object

Instead of using a boolean, just pass the object directly into the function

* Add changelog

* Refactor tests into separate cases

* Simplify if-condition

This is will not raise a nil-pointer exception because the `ConnectionConfig` field is never dereferenced if the object it is called on is `nil`:

    func (x *Upstream) GetConnectionConfig() *ConnectionConfig {
      if x != nil {
        return x.ConnectionConfig
      }
      return nil
    }

Co-authored-by: soloio-bulldozer[bot] <48420018+soloio-bulldozer[bot]@users.noreply.github.com>

* Move changelog into v1.11

Co-authored-by: soloio-bulldozer[bot] <48420018+soloio-bulldozer[bot]@users.noreply.github.com>