Initial documentation on LDAP auth with service account
This commit adds some initial documentation on how to authenticate against LDAP with a service account. I have only included what I believe to be the contents necessary to configure this, but have not tested it in too much depth or proofread for language. I would appreciate if the documentation team can finish it off from here.