prevent Proxy namespace from being modified
Previously, a 'metaKey' was generated for each Proxy being translated
and synced to envoy. This key was used as a 'tag' for tracing of the
translation/sync prcoess. It used a custom function that was almost exactly
matched how the SnapshotCacheKey used for xDS was generated; this function
also modified the actual Proxy's namespace, which broke assumptions of where
the proxy lives in the in-memory client.

By removing the namespace modification we fix various status error logs as
now the Proxy is located where the rest of the system expects (the write
namespace). Also, we can simply use the SnapshotCacheKey(...) to generate
a key with the same logic used for xds.