[1.17] fix: allows setting Istio CA_ADDR independently (#9897)
* fix: allows setting Istio CA_ADDR independently

https://istio.io/latest/docs/reference/commands/pilot-agent/

The Istio documentation indicates that CA_ADDR defaults to the
PROXY_CONFIG discovery address. Though it should be possible to specify
a CA_ADDR that is unrelated to the discovery address.

However, Gloo helm forces these two separate fields to be aligned as it
is driven from a single helm value.

This change introduces an `istioSpiffeCertProviderAddress` property
resolving the problem. The default is kept as `istioDiscoveryAddress` to
ensure backward compatibility.

* docs: add property documentation + issue link

* move changelog to beta15 folder

* helm_test: fix compile issues, resolve some deprecated function calls

---------

Co-authored-by: Jérémy Lourenço <jeremy.lourenco@snyk.io>
Co-authored-by: soloio-bulldozer[bot] <48420018+soloio-bulldozer[bot]@users.noreply.github.com>