test fix: allow test scenarios run more than one time
Setting the logger in controller runtime only works for the first time, and is no-op the second time. this means that the second time the test runs it will run to a testing.T that finished running.

The fix is to changing the underlying log writer, so we can swap up the testing.T to the current one