gateway2/delegation: fix cycle detection
There's a bug in how cycles are detected when evaluating a
delegation chain, because of which a multi-level delegation
tree with the same child route being referenced by multiple
parent HTTP routes is broken. This change fixes the detection
of cycles during the evaluation of the delegation chain.

---
Testing done:

Added a new translator unit test which fails without the fix.

Signed-off-by: Shashank Ram <shashank.ram@solo.io>