gateway2: allow route delegation using well known label
There is a product requirement to enable users to use
a label to select HTTPRoutes to delegate to instead
of GVK ref to other HTTPRoutes (includes wildcards).

To strike a balance between flexibility and performance,
this change implements the proposal to use a well known
label `delegation.gateway.solo.io/label=<value>` to
allow users to delegate to other HTTPRoutes using a label.
HTTPRoutes are indexed using this well known label key that
enable O(1) lookups of routes matching this label value.

Also moves user-facing labels and annotations to the wellknown
pkg to consolidate them for visibility.

Signed-off-by: Shashank Ram <shashank.ram@solo.io>