gateway2/status: correctly merge parentRefs for routes
When a route has multiple parentRefs, the exitsing status
merging code does not correctly merge parentRefs belonging
to different proxies. As a result, a route with multiple
parents does not have both parents in their status during
merging of status reports.
This change fixes it.

It also ensures we do not write parentRef statuses
that the reporter does not own.

Signed-off-by: Shashank Ram <shashank.ram@solo.io>